完善下载进度显示

+++ app/src/main/java/com/hiroshi/cimoc/presenter/DetailPresenter.java
@@ -139,9 +139,27 @@ public class DetailPresenter extends BasePresenter<DetailView> {
                 });
     }
 
-    public long addDownload(long key, String path) {
-        Task task = new Task(null, key, path, 0, false);
-        return mTaskManager.insert(task);
+    public Task addTask(String path, String title) {
+        Task task = new Task(null, -1, path, title, 0, 0, false);
+
+        Long key = mComic.getId();
+        mComic.setDownload(System.currentTimeMillis());
+        if (key != null) {
+            mComicManager.update(mComic);
+            task.setKey(key);
+        } else {
+            long id = mComicManager.insert(mComic);
+            mComic.setId(id);
+            task.setKey(id);
+        }
+
+        long id = mTaskManager.insert(task);
+        task.setId(id);
+        task.setInfo(mComic.getSource(), mComic.getCid(), mComic.getTitle());
+        task.setState(Task.STATE_WAIT);
+
+        RxBus.getInstance().post(new RxEvent(RxEvent.TASK_ADD, task));
+        return task;
     }
 
     public void updateComic() {
@@ -170,7 +188,7 @@ public class DetailPresenter extends BasePresenter<DetailView> {
     public void unfavoriteComic() {
         long id = mComic.getId();
         mComic.setFavorite(null);
-        if (mComic.getHistory() == null) {
+        if (mComic.getHistory() == null && mComic.getDownload() == null) {
             mComicManager.deleteByKey(id);
             mComic.setId(null);
         }
